requirements.txt

# third party packages
dvc
dvc[gdrive]
scikit-learn
pandas
pytest
tox
flake8
flask
gunicorn
PyYAML
from_root